Revert changes to utils
