+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/bestpractices/PreserveStackTraceRule.java
@@ -27,7 +27,6 @@ import net.sourceforge.pmd.lang.java.ast.ASTVariableDeclaratorId;
 import net.sourceforge.pmd.lang.java.ast.InvocationNode;
 import net.sourceforge.pmd.lang.java.ast.JavaNode;
 import net.sourceforge.pmd.lang.java.rule.AbstractJavaRulechainRule;
-import net.sourceforge.pmd.lang.java.rule.internal.JavaAstUtil;
 import net.sourceforge.pmd.lang.java.rule.internal.JavaRuleUtil;
 import net.sourceforge.pmd.lang.java.symbols.JVariableSymbol;
 import net.sourceforge.pmd.lang.java.types.InvocationMatcher;
@@ -83,7 +82,7 @@ public class PreserveStackTraceRule extends AbstractJavaRulechainRule {
 
         } else if (expr instanceof ASTCastExpression) {
 
-            ASTExpression innermost = JavaAstUtil.peelCasts(expr);
+            ASTExpression innermost = JavaRuleUtil.peelCasts(expr);
             return exprConsumesException(exceptionParam, innermost, mayBeSelf);
 
         } else if (expr instanceof ASTConditionalExpression) {
@@ -122,7 +121,7 @@ public class PreserveStackTraceRule extends AbstractJavaRulechainRule {
                     return true;
                 }
 
-                if (JavaAstUtil.followingCallChain(usage).any(it -> consumesExceptionNonRecursive(exceptionParam, it))) {
+                if (JavaRuleUtil.followingCallChain(usage).any(it -> consumesExceptionNonRecursive(exceptionParam, it))) {
                     return true;
                 }
             }
@@ -136,12 +135,12 @@ public class PreserveStackTraceRule extends AbstractJavaRulechainRule {
 
     private boolean assignmentRhsConsumesException(ASTVariableDeclaratorId exceptionParam, ASTVariableDeclaratorId lhsVariable, ASTNamedReferenceExpr usage) {
         if (usage.getIndexInParent() == 0) {
-            ASTExpression assignmentRhs = JavaAstUtil.getOtherOperandIfInAssignmentExpr(usage);
+            ASTExpression assignmentRhs = JavaRuleUtil.getOtherOperandIfInAssignmentExpr(usage);
             boolean rhsIsSelfReferential =
                 NodeStream.of(assignmentRhs)
                           .descendantsOrSelf()
                           .filterIs(ASTVariableAccess.class)
-                          .any(it -> JavaAstUtil.isReferenceToVar(it, lhsVariable.getSymbol()));
+                          .any(it -> JavaRuleUtil.isReferenceToVar(it, lhsVariable.getSymbol()));
             return !rhsIsSelfReferential && exprConsumesException(exceptionParam, assignmentRhs, true);
         }
         return false;
