+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/design/ImmutableFieldRule.java
@@ -24,8 +24,8 @@ import net.sourceforge.pmd.lang.java.rule.AbstractJavaRulechainRule;
 import net.sourceforge.pmd.lang.java.rule.internal.DataflowPass;
 import net.sourceforge.pmd.lang.java.rule.internal.DataflowPass.AssignmentEntry;
 import net.sourceforge.pmd.lang.java.rule.internal.DataflowPass.DataflowResult;
-import net.sourceforge.pmd.lang.java.rule.internal.JavaAstUtil;
 import net.sourceforge.pmd.lang.java.rule.internal.JavaPropertyUtil;
+import net.sourceforge.pmd.lang.java.rule.internal.JavaRuleUtil;
 import net.sourceforge.pmd.properties.PropertyDescriptor;
 import net.sourceforge.pmd.util.CollectionUtil;
 
@@ -58,9 +58,9 @@ public class ImmutableFieldRule extends AbstractJavaRulechainRule {
         ASTAnyTypeDeclaration enclosingType = field.getEnclosingType();
         if (field.getEffectiveVisibility().isAtMost(Visibility.V_PRIVATE)
             && !field.getModifiers().hasAny(JModifier.VOLATILE, JModifier.STATIC, JModifier.FINAL)
-            && !JavaAstUtil.hasAnyAnnotation(field, INVALIDATING_FIELD_ANNOTS)
-            && !JavaAstUtil.hasAnyAnnotation(enclosingType, INVALIDATING_CLASS_ANNOT)
-            && !JavaAstUtil.hasAnyAnnotation(field, getProperty(IGNORED_ANNOTS))) {
+            && !JavaRuleUtil.hasAnyAnnotation(field, INVALIDATING_FIELD_ANNOTS)
+            && !JavaRuleUtil.hasAnyAnnotation(enclosingType, INVALIDATING_CLASS_ANNOT)
+            && !JavaRuleUtil.hasAnyAnnotation(field, getProperty(IGNORED_ANNOTS))) {
 
             DataflowResult dataflow = DataflowPass.getDataflowResult(field.getRoot());
