+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/errorprone/IdempotentOperationsRule.java
@@ -7,7 +7,7 @@ package net.sourceforge.pmd.lang.java.rule.errorprone;
 import net.sourceforge.pmd.lang.java.ast.ASTAssignmentExpression;
 import net.sourceforge.pmd.lang.java.ast.AssignmentOp;
 import net.sourceforge.pmd.lang.java.rule.AbstractJavaRulechainRule;
-import net.sourceforge.pmd.lang.java.rule.internal.JavaAstUtil;
+import net.sourceforge.pmd.lang.java.rule.internal.JavaRuleUtil;
 
 public class IdempotentOperationsRule extends AbstractJavaRulechainRule {
 
@@ -18,7 +18,7 @@ public class IdempotentOperationsRule extends AbstractJavaRulechainRule {
     @Override
     public Object visit(ASTAssignmentExpression node, Object data) {
         if (node.getOperator() == AssignmentOp.ASSIGN
-            && JavaAstUtil.isReferenceToSameVar(node.getLeftOperand(), node.getRightOperand())) {
+            && JavaRuleUtil.isReferenceToSameVar(node.getLeftOperand(), node.getRightOperand())) {
             addViolation(data, node);
         }
         return null;
