+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/performance/UseStringBufferForStringAppendsRule.java
@@ -14,7 +14,7 @@ import net.sourceforge.pmd.lang.java.ast.ASTAssignmentExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTLoopStatement;
 import net.sourceforge.pmd.lang.java.ast.ASTVariableDeclaratorId;
 import net.sourceforge.pmd.lang.java.rule.AbstractJavaRulechainRule;
-import net.sourceforge.pmd.lang.java.rule.internal.JavaAstUtil;
+import net.sourceforge.pmd.lang.java.rule.internal.JavaRuleUtil;
 import net.sourceforge.pmd.lang.java.types.TypeTestUtil;
 
 public class UseStringBufferForStringAppendsRule extends AbstractJavaRulechainRule {
@@ -56,10 +56,10 @@ public class UseStringBufferForStringAppendsRule extends AbstractJavaRulechainRu
                 }
 
                 int usageOnRightHandSide =
-                    JavaAstUtil.flattenOperands(assignment.getRightOperand())
-                               .filterIs(ASTNamedReferenceExpr.class)
-                               .filterMatching(ASTNamedReferenceExpr::getReferencedSym, node.getSymbol())
-                               .count();
+                    JavaRuleUtil.flattenOperands(assignment.getRightOperand())
+                                .filterIs(ASTNamedReferenceExpr.class)
+                                .filterMatching(ASTNamedReferenceExpr::getReferencedSym, node.getSymbol())
+                                .count();
 
                 // or maybe a append in some way (a = a + x)
                 // or a combination (a += a + x)
