+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/performance/UselessStringValueOfRule.java
@@ -10,7 +10,7 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 import net.sourceforge.pmd.lang.java.ast.ASTExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodCall;
 import net.sourceforge.pmd.lang.java.rule.AbstractJavaRule;
-import net.sourceforge.pmd.lang.java.rule.internal.JavaAstUtil;
+import net.sourceforge.pmd.lang.java.rule.internal.JavaRuleUtil;
 import net.sourceforge.pmd.lang.java.types.TypeTestUtil;
 import net.sourceforge.pmd.lang.rule.RuleTargetSelector;
 
@@ -24,7 +24,7 @@ public class UselessStringValueOfRule extends AbstractJavaRule {
 
     @Override
     public Object visit(ASTMethodCall node, Object data) {
-        if (JavaAstUtil.isStringConcatExpr(node.getParent())) {
+        if (JavaRuleUtil.isStringConcatExpr(node.getParent())) {
             ASTExpression valueOfArg = getValueOfArg(node);
             if (valueOfArg == null) {
                 return data; //not a valueOf call
@@ -33,7 +33,7 @@ public class UselessStringValueOfRule extends AbstractJavaRule {
                 return data;
             }
 
-            ASTExpression sibling = JavaAstUtil.getOtherOperandIfInInfixExpr(node);
+            ASTExpression sibling = JavaRuleUtil.getOtherOperandIfInInfixExpr(node);
             if (TypeTestUtil.isExactlyA(String.class, sibling)
                 && !valueOfArg.getTypeMirror().isArray()
                 // In `String.valueOf(a) + String.valueOf(b)`,
