+++ pmd-java/src/test/kotlin/net/sourceforge/pmd/lang/java/types/ast/ConversionContextTests.kt
@@ -10,7 +10,7 @@ import io.kotest.matchers.shouldBe
 import net.sourceforge.pmd.lang.ast.test.component6
 import net.sourceforge.pmd.lang.ast.test.shouldBe
 import net.sourceforge.pmd.lang.java.ast.*
-import net.sourceforge.pmd.lang.java.rule.internal.JavaAstUtil
+import net.sourceforge.pmd.lang.java.rule.internal.JavaRuleUtil
 import net.sourceforge.pmd.lang.java.types.STRING
 import net.sourceforge.pmd.lang.java.types.parseWithTypeInferenceSpy
 import net.sourceforge.pmd.lang.java.types.shouldHaveType
@@ -239,9 +239,7 @@ class ConversionContextTests : ProcessorTestSpec({
         spy.shouldBeOk {
             concats.forEach {
                 withClue(it) {
-                    JavaAstUtil.isStringConcatExpr(
-                        it
-                    ) shouldBe true
+                    JavaRuleUtil.isStringConcatExpr(it) shouldBe true
                     it.leftOperand.conversionContext::getTargetType shouldBe ts.STRING
                     it.rightOperand.conversionContext::getTargetType shouldBe ts.STRING
                 }
